Remove unneeded exception
As os.uname will not lead to an exception itself as the python documentation does tell when a funtion in it raises an exception. And as we can eliminate the need to index directly using the get method there is no need to raise and log an exception, when we can get much more concise and readable code.