Rearchitected the package manager to be performant
As it stands the package manager was creating a subprocess which called
another subprocess within it. As creating a new subprocess isn't exactly
cheap, I think it is better to use and if __name__ == "__main__" guard in
Actions.py and import it directly and use it as module when needed and
as a invokable command when needed.